refactor: Add comments data and route
Add the comments data file and route to the project. This includes the creation of the `data.ts` file with an array of comments and the `route.ts` file with GET and POST functions for retrieving and adding comments. These changes enable the functionality to display and submit comments in the application.